# docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
...
  choices.AgeType:
    enum:
    - 4
    - 12
    - 16
    - 18
    type: integer
    x-enum-varnames:
    - ATYPE_FOUR
    - ATYPE_TWELVE
    - ATYPE_SIXTEEN
    - ATYPE_EIGHTEEN
  choices.ChapterStatus:
    enum:
    - DRAFT
    - PUBLISHED
    - TRASH
    type: string
    x-enum-varnames:
    - CS_DRAFT
    - CS_PUBLISHED
    - CS_TRASH
  choices.ContractIDTypeChoice:
    enum:
    - DRIVERS-LICENSE
    - GOVERNMENT-ID
    - PASSPORT
    type: string
    x-enum-varnames:
    - CID_DRIVERS_LICENSE
    - CID_GOVERNMENT_ID
    - CID_PASSPORT
  choices.ContractStatusChoice:
    enum:
    - PENDING
    - APPROVED
    - DECLINED
    - UPDATED
    type: string
    x-enum-varnames:
    - CTS_PENDING
    - CTS_APPROVED
    - CTS_DECLINED
    - CTS_UPDATED
  choices.ContractTypeChoice:
    enum:
    - EXCLUSIVE
    - NON-EXCLUSIVE
    - ONLY-EXCLUSIVE
    type: string
    x-enum-varnames:
    - CT_EXCLUSIVE
    - CT_NON_EXCLUSIVE
    - CT_ONLY_EXCLUSIVE
...
  choices.RatingChoice:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    type: integer
    x-enum-varnames:
    - RC_1
    - RC_2
    - RC_3
    - RC_4
    - RC_5
...
  schemas.BookDetailResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.BookDetailSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
  schemas.BookDetailSchema:
    properties:
      age_discretion:
        $ref: '#/definitions/choices.AgeType'
      author:
        $ref: '#/definitions/schemas.UserDataSchema'
      avg_rating:
        type: number
      blurb:
        type: string
      chapter_price:
        type: integer
      chapters_count:
        type: integer
      cover_image:
        type: string
      created_at:
        example: "2024-06-05T02:32:34.462196+01:00"
        type: string
      full_price:
        type: integer
      genre:
        $ref: '#/definitions/schemas.GenreWithoutTagSchema'
      partial_view_chapter:
        $ref: '#/definitions/schemas.ChapterSchema'
      reviews:
        $ref: '#/definitions/schemas.ReviewsResponseDataSchema'
      slug:
        type: string
      tags:
        items:
          $ref: '#/definitions/schemas.TagSchema'
        type: array
      title:
        type: string
      updated_at:
        example: "2024-06-05T02:32:34.462196+01:00"
        type: string
      views:
        type: integer
      votes:
        type: integer
      word_count:
        type: integer
    type: object
...
  schemas.BookSchema:
    properties:
      age_discretion:
        $ref: '#/definitions/choices.AgeType'
      author:
        $ref: '#/definitions/schemas.UserDataSchema'
      avg_rating:
        type: number
      blurb:
        type: string
      chapter_price:
        type: integer
      chapters_count:
        type: integer
      cover_image:
        type: string
      created_at:
        example: "2024-06-05T02:32:34.462196+01:00"
        type: string
      full_price:
        type: integer
      genre:
        $ref: '#/definitions/schemas.GenreWithoutTagSchema'
      partial_view_chapter:
        $ref: '#/definitions/schemas.ChapterSchema'
      slug:
        type: string
      tags:
        items:
          $ref: '#/definitions/schemas.TagSchema'
        type: array
      title:
        type: string
      updated_at:
        example: "2024-06-05T02:32:34.462196+01:00"
        type: string
      views:
        type: integer
      votes:
        type: integer
      word_count:
        type: integer
    type: object
...
  schemas.BooksResponseDataSchema:
    properties:
      books:
        items:
          $ref: '#/definitions/schemas.BookSchema'
        type: array
      current_page:
        example: 1
        type: integer
      last_page:
        example: 100
        type: integer
      per_page:
        example: 100
        type: integer
    type: object
  schemas.BooksResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.BooksResponseDataSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  schemas.ChapterSchema:
    properties:
      chapter_status:
        allOf:
        - $ref: '#/definitions/choices.ChapterStatus'
        example: PUBLISHED
      slug:
        type: string
      text:
        type: string
      title:
        type: string
      word_count:
        type: integer
    type: object
...
  schemas.ContractSchema:
    properties:
      address:
        type: string
      age:
        type: integer
      average_chapter:
        type: integer
      book_availability_link:
        type: string
      chapter_price:
        type: integer
      city:
        type: string
      contract_status:
        $ref: '#/definitions/choices.ContractStatusChoice'
      country:
        type: string
      email:
        type: string
      full_name:
        type: string
      full_price:
        type: integer
      full_purchase_mode:
        type: boolean
      id_back_image:
        type: string
      id_front_image:
        type: string
      id_type:
        $ref: '#/definitions/choices.ContractIDTypeChoice'
      intended_contract:
        $ref: '#/definitions/choices.ContractTypeChoice'
      outline:
        type: string
      pen_name:
        type: string
      planned_length:
        type: integer
      postal_code:
        type: integer
      state:
        type: string
      synopsis:
        type: string
      telephone_number:
        type: string
      update_rate:
        type: integer
    type: object
  schemas.ContractsResponseDataSchema:
    properties:
      contracts:
        items:
          $ref: '#/definitions/schemas.ContractSchema'
        type: array
      current_page:
        example: 1
        type: integer
      last_page:
        example: 100
        type: integer
      per_page:
        example: 100
        type: integer
    type: object
  schemas.ContractsResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.ContractsResponseDataSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  schemas.GenreAddSchema:
    properties:
      name:
        type: string
      tag_slugs:
        items:
          type: string
        type: array
    required:
    - name
    type: object
...
  schemas.GenreWithoutTagSchema:
    properties:
      name:
        type: string
      slug:
        type: string
    type: object
...
  schemas.ResponseSchema:
    properties:
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  schemas.ReviewSchema:
    properties:
      created_at:
        example: "2024-06-05T02:32:34.462196+01:00"
        type: string
      id:
        example: 2b3bd817-135e-41bd-9781-33807c92ff40
        type: string
      likes_count:
        type: integer
      rating:
        $ref: '#/definitions/choices.RatingChoice'
      replies_count:
        type: integer
      text:
        maxLength: 10000
        type: string
      updated_at:
        example: "2024-06-05T02:32:34.462196+01:00"
        type: string
      user:
        $ref: '#/definitions/schemas.UserDataSchema'
    required:
    - rating
    - text
    type: object
  schemas.ReviewsResponseDataSchema:
    properties:
      current_page:
        example: 1
        type: integer
      items:
        items:
          $ref: '#/definitions/schemas.ReviewSchema'
        type: array
      last_page:
        example: 100
        type: integer
      per_page:
        example: 100
        type: integer
    type: object
...
  schemas.TagSchema:
    properties:
      name:
        type: string
      slug:
        type: string
    type: object
  schemas.TagsAddSchema:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  schemas.UserDataSchema:
    properties:
      avatar:
        type: string
      name:
        description: For short user data
        type: string
      username:
        type: string
    type: object
...
  utils.ErrorResponse:
    properties:
      code:
        type: string
      data:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: |-
    ## LitPAD API built with Fiber and GORM

    ### WEBSOCKETS:

    #### Notifications

    - URL: `wss://{host}/api/v1/ws/notifications`

    - Requires authorization, so pass in the Bearer Authorization header.

    - You can only read and not send notification messages into this socket.
  title: LITPAD API
  version: "1.0"
paths:
...
  /admin/books:
    get:
      consumes:
      - application/json
      description: Retrieves a list of books with support for pagination and optional
        filtering based on book title.
      parameters:
      - default: 1
        description: Current Page
        in: query
        name: page
        type: integer
      - description: Title of the book to filter by
        in: query
        name: title
        type: string
      - description: name or username of the book author to filter by
        in: query
        name: name
        type: string
      - description: Filter by highest ratings
        in: query
        name: rating
        type: boolean
      - description: Filter by Genre slug
        in: query
        name: genre_slug
        type: string
      - description: Filter by Tag slug
        in: query
        name: tag_slug
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of books
          schema:
            $ref: '#/definitions/schemas.BooksResponseSchema'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Books with Pagination
      tags:
      - Admin | Books
  /admin/books/book-detail/{slug}:
    get:
      description: This endpoint allows an admin to view details of a book
      parameters:
      - default: 1
        description: Current Page (for reviews pagination)
        in: query
        name: page
        type: integer
      - description: Book slug
        in: path
        name: slug
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.BookDetailResponseSchema'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: View Book Details
      tags:
      - Admin | Books
  /admin/books/by-username/{username}:
    get:
      consumes:
      - application/json
      description: Retrieves a list of a particular author books with support for
        pagination and optional filtering based on book title.
      parameters:
      - description: Username of the author
        in: path
        name: username
        required: true
        type: string
      - default: 1
        description: Current Page
        in: query
        name: page
        type: integer
      - description: Title of the book to filter by
        in: query
        name: title
        type: string
      - description: Filter by highest ratings
        in: query
        name: rating
        type: boolean
      - description: Filter by Genre slug
        in: query
        name: genre_slug
        type: string
      - description: Filter by Tag slug
        in: query
        name: tag_slug
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of books
          schema:
            $ref: '#/definitions/schemas.BooksResponseSchema'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Author Books with Pagination
      tags:
      - Admin | Books
  /admin/books/contracts:
    get:
      consumes:
      - application/json
      description: Retrieves a list of book contracts with support for pagination
        and optional filtering based on contract status.
      parameters:
      - default: 1
        description: Current Page
        in: query
        name: page
        type: integer
      - description: Name of the author to filter by
        in: query
        name: name
        type: string
      - description: status of the contract to filter by
        enum:
        - PENDING
        - APPROVED
        - DECLINED
        - UPDATED
        in: query
        name: contract_status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of book contracts
          schema:
            $ref: '#/definitions/schemas.ContractsResponseSchema'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Book Contracts with Pagination
      tags:
      - Admin | Books
  /admin/books/genres:
    post:
      consumes:
      - application/json
      description: Add a new genre to the app.
      parameters:
      - description: Genre
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schemas.GenreAddSchema'
      produces:
      - application/json
      responses:
        "201":
          description: Genre Added Successfully
          schema:
            $ref: '#/definitions/schemas.ResponseSchema'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add Genre
      tags:
      - Admin | Books
  /admin/books/genres/{slug}:
    delete:
      consumes:
      - application/json
      description: Delete a genre.
      parameters:
      - description: Genre slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Genre Deleted Successfully
          schema:
            $ref: '#/definitions/schemas.ResponseSchema'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Genre
      tags:
      - Admin | Books
    put:
      consumes:
      - application/json
      description: Update a genre.
      parameters:
      - description: Genre slug
        in: path
        name: slug
        required: true
        type: string
      - description: Genre
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schemas.GenreAddSchema'
      produces:
      - application/json
      responses:
        "200":
          description: Genre Updated Successfully
          schema:
            $ref: '#/definitions/schemas.ResponseSchema'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Genre
      tags:
      - Admin | Books
  /admin/books/tags:
    post:
      consumes:
      - application/json
      description: Add a new tag to the app.
      parameters:
      - description: Tag
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schemas.TagsAddSchema'
      produces:
      - application/json
      responses:
        "201":
          description: Tag added successfully
          schema:
            $ref: '#/definitions/schemas.ResponseSchema'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add Tag
      tags:
      - Admin | Books
  /admin/books/tags/{slug}:
    delete:
      consumes:
      - application/json
      description: Delete a tag from the app.
      parameters:
      - description: Tag slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tag delete successfully
          schema:
            $ref: '#/definitions/schemas.ResponseSchema'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Tag
      tags:
      - Admin | Books
    put:
      consumes:
      - application/json
      description: Update a tag to the app.
      parameters:
      - description: Tag slug
        in: path
        name: slug
        required: true
        type: string
      - description: Tag
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schemas.TagsAddSchema'
      produces:
      - application/json
      responses:
        "200":
          description: Tag updated successfully
          schema:
            $ref: '#/definitions/schemas.ResponseSchema'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Tag
      tags:
      - Admin | Books
...
produces:
- application/json
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer jwt_string' to correctly set the API Key
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"